requirements.txt
black==26.3.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.2
idna==3.12
mypy_extensions==1.1.0
packaging==26.1
pathspec==1.0.4
platformdirs==4.9.6
python-dotenv==1.2.2
pytokens==0.4.1
requests==2.33.1
urllib3==2.6.3
